+++ src/QueryBuilder/AbstractQueryBuilder.php
@@ -197,6 +197,11 @@ abstract class AbstractQueryBuilder implements QueryBuilderInterface
         return $this->dqlBuilder->buildExpression($expression, $params);
     }
 
+    public function buildFor(array $values): string
+    {
+        return $this->dqlBuilder->buildFor($values);
+    }
+
     public function buildFrom(array|null $tables, array &$params): string
     {
         return $this->dqlBuilder->buildFrom($tables, $params);