+++ tests/AbstractQueryBuilderTest.php
@@ -33,6 +33,9 @@ use Yiisoft\Db\Tests\Support\Assert;
 use Yiisoft\Db\Tests\Support\DbHelper;
 use Yiisoft\Db\Tests\Support\TestTrait;
 
+use function PHPUnit\Framework\assertEmpty;
+use function PHPUnit\Framework\assertSame;
+
 /**
  * @psalm-suppress PropertyNotSetInConstructor
  */
@@ -349,11 +352,39 @@ abstract class AbstractQueryBuilderTest extends TestCase
         $this->assertSame($expectedParams, $params);
     }
 
-    /**
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws NotSupportedException
-     */
+    public static function dataBuildFor(): iterable
+    {
+        yield ['', []];
+        yield ['FOR UPDATE', ['UPDATE']];
+        yield ['FOR UPDATE FOR SHARE', ['UPDATE', 'SHARE']];
+    }
+
+    #[DataProvider('dataBuildFor')]
+    public function testBuildFor(string $expected, array $value): void
+    {
+        $queryBuilder = $this->getConnection()->getQueryBuilder();
+        assertSame($expected, $queryBuilder->buildFor($value));
+    }
+
+    public function testBuildWithFor(): void
+    {
+        $db = $this->getConnection();
+        $queryBuilder = $db->getQueryBuilder();
+
+        $query = (new Query($db))->from('test')->for('UPDATE OF {{t1}}');
+
+        [$sql, $params] = $queryBuilder->build($query);
+
+        assertSame(
+            DbHelper::replaceQuotes(
+                'SELECT * FROM [[test]] FOR UPDATE OF {{t1}}',
+                $db->getDriverName(),
+            ),
+            $sql
+        );
+        assertEmpty($params);
+    }
+
     public function testBuildFrom(): void
     {
         $db = $this->getConnection();