+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 - Enh #963: Make `Query::andHaving()` similar to `Query::andWhere()` (@Tigrov)
 - New #964: Add `QueryBuilderInterface::replacePlaceholders()` method (@Tigrov)
 - Enh #879: Rename `getLastInsertID()` method in `ConnectionInterface` to `getLastInsertId()` (@vjik)
+- New #967: Add `FOR` clause to query (@vjik)
 
 ## 1.3.0 March 21, 2024